refactor(product): dto 경로 변경, 공유 dto에서 특정 도메인 의존성 삭제
