+++ src/main/java/com/bugzero/rarego/boundedContext/auction/app/AuctionCreateAuctionUseCase.java
@@ -2,6 +2,7 @@ package com.bugzero.rarego.boundedContext.auction.app;
 
 import org.springframework.stereotype.Service;
 
+import com.bugzero.rarego.boundedContext.auction.domain.Auction;
 import com.bugzero.rarego.boundedContext.auction.domain.AuctionMember;
 import com.bugzero.rarego.boundedContext.auction.out.AuctionRepository;
 import com.bugzero.rarego.shared.product.dto.ProductAuctionRequestDto;
@@ -15,11 +16,18 @@ public class AuctionCreateAuctionUseCase {
 	private final AuctionSupport auctionSupport;
 
 	// 신규상품 경매 정보 생성
-	public long createAuction(Long productId,String publicId, ProductAuctionRequestDto productAuctionRequestDto) {
+	public long createAuction(Long productId,String publicId, ProductAuctionRequestDto dto) {
 		AuctionMember seller = auctionSupport.getPublicMember(publicId);
 
 		return auctionRepository
-			.save(productAuctionRequestDto.toEntity(productId, seller.getId()))
+			.save(Auction.builder()
+				.productId(productId)
+				.sellerId(seller.getId())
+				.startPrice(dto.startPrice())
+				.durationDays(dto.durationDays())
+				.build())
 			.getId();
+
+
 	}
 }
