+++ src/main/java/com/bugzero/rarego/boundedContext/product/app/ProductReadInspectionUseCase.java
@@ -4,10 +4,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.bugzero.rarego.boundedContext.product.domain.Inspection;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductInspectionResponseDto;
 import com.bugzero.rarego.boundedContext.product.out.InspectionRepository;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.response.ErrorType;
-import com.bugzero.rarego.shared.product.dto.ProductInspectionResponseDto;
 
 import lombok.RequiredArgsConstructor;
