+++ src/main/java/com/bugzero/rarego/boundedContext/product/app/ProductUpdateProductUseCase.java
@@ -7,10 +7,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.bugzero.rarego.boundedContext.product.domain.Product;
 import com.bugzero.rarego.boundedContext.product.domain.ProductMember;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductImageUpdateDto;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductUpdateResponseDto;
 import com.bugzero.rarego.shared.auction.out.AuctionApiClient;
-import com.bugzero.rarego.shared.product.dto.ProductImageUpdateDto;
 import com.bugzero.rarego.shared.product.dto.ProductUpdateDto;
-import com.bugzero.rarego.shared.product.dto.ProductUpdateResponseDto;
 
 import lombok.RequiredArgsConstructor;
