+++ src/main/java/com/bugzero/rarego/shared/product/dto/ProductUpdateDto.java
@@ -3,6 +3,7 @@ package com.bugzero.rarego.shared.product.dto;
 import java.util.List;
 
 import com.bugzero.rarego.boundedContext.product.domain.Category;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductImageUpdateDto;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
