+++ src/test/java/com/bugzero/rarego/boundedContext/auction/app/AuctionCreateAuctionUseCaseTest.java
@@ -51,7 +51,14 @@ class AuctionCreateAuctionUseCaseTest {
 			.build();
 
 		// Mocking: 저장 시 ID가 100인 객체가 반환된다고 가정
-		Auction mockAuction = requestDto.toEntity(productId, 1L);
+		Auction mockAuction = Auction.builder()
+			.productId(productId)
+			.sellerId(commonSeller.getId())
+			.startPrice(requestDto.startPrice())
+			.durationDays(requestDto.durationDays())
+			.build();
+
+
 		ReflectionTestUtils.setField(mockAuction, "id", 100L);
 		when(auctionRepository.save(any(Auction.class))).thenReturn(mockAuction);
 		given(auctionSupport.getPublicMember(PUBLIC_ID)).willReturn(commonSeller);
