+++ src/test/java/com/bugzero/rarego/boundedContext/product/app/ProductCreateInspectionUseCaseTest.java
@@ -20,11 +20,11 @@ import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
 import com.bugzero.rarego.boundedContext.product.domain.Product;
 import com.bugzero.rarego.boundedContext.product.domain.ProductCondition;
 import com.bugzero.rarego.boundedContext.product.domain.ProductMember;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductInspectionRequestDto;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductInspectionResponseDto;
 import com.bugzero.rarego.boundedContext.product.out.InspectionRepository;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.response.ErrorType;
-import com.bugzero.rarego.shared.product.dto.ProductInspectionRequestDto;
-import com.bugzero.rarego.shared.product.dto.ProductInspectionResponseDto;
 
 @ExtendWith(MockitoExtension.class)
 class ProductCreateInspectionUseCaseTest {
