+++ src/test/java/com/bugzero/rarego/boundedContext/product/app/ProductReadInspectionUseCaseTest.java
@@ -18,10 +18,10 @@ import org.springframework.test.util.ReflectionTestUtils;
 import com.bugzero.rarego.boundedContext.product.domain.Inspection;
 import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
 import com.bugzero.rarego.boundedContext.product.domain.ProductCondition;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductInspectionResponseDto;
 import com.bugzero.rarego.boundedContext.product.out.InspectionRepository;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.response.ErrorType;
-import com.bugzero.rarego.shared.product.dto.ProductInspectionResponseDto;
 
 @ExtendWith(MockitoExtension.class)
 class ProductReadInspectionUseCaseTest {
