+++ src/test/java/com/bugzero/rarego/boundedContext/product/app/ProductReadProductsForInspectionUseCaseTest.java
@@ -18,8 +18,8 @@ import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
 import com.bugzero.rarego.boundedContext.product.domain.Product;
 import com.bugzero.rarego.boundedContext.product.domain.ProductImage;
 import com.bugzero.rarego.boundedContext.product.domain.ProductMember;
+import com.bugzero.rarego.boundedContext.product.domain.dto.ProductResponseForInspectionDto;
 import com.bugzero.rarego.boundedContext.product.out.ProductRepository;
-import com.bugzero.rarego.shared.product.dto.ProductResponseForInspectionDto;
 
 @DataJpaTest
 class ProductReadProductsForInspectionUseCaseTest {
